Extracted generic test library code.
